requirements.txt
Pillow
streamlink
tesserocr
opencv2